requirements.txt
aioredis==1.2.0
aniso8601==6.0.0
argon2-cffi==19.1.0
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
automium==0.2.6
automium-web==0.1.1
awesome-slugify==1.6.5
boto3==1.9.202
botocore==1.12.202
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.0.13
django-allauth==0.39.1
django-anymail==6.1.0
django-contrib-comments==1.9.1
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==2.0.2
django-jquery==3.1.0
django-markdownx==2.0.28
django-oauth-toolkit==1.2.0
django-redis==4.10.0
django-rest-auth==0.9.5
django-storages==1.7.1
django-taggit==1.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
docutils==0.14
env==0.1.0
et-xmlfile==1.0.1
Flask==1.1.2
gevent==1.4.0
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.15
gunicorn==19.9.0
heroku==0.1.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.9.4
legacy==0.1.6
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==0.6.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==6.1.0
postgres==2.2.2
promise==2.2.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
pygame==1.9.6
PyHamcrest==1.9.0
python-dateutil==2.8.0
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.3
raven==6.10.0
redis==3.3.6
regex==2019.6.8
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
tablib==1.1.0
Twisted==19.7.0
txaio==18.8.1
Unidecode==0.4.21
urllib3==1.25.3
views==0.3
virtualenv==16.7.1
virtualenvwrapper-win==1.2.5
Werkzeug==1.0.1
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
